@@ -29,39 +29,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "16.0.0",
-    "@angular/cdk": "16.0.0",
-    "@angular/common": "16.0.0",
-    "@angular/compiler": "16.0.0",
-    "@angular/core": "16.0.0",
-    "@angular/forms": "16.0.0",
-    "@angular/material": "16.0.0",
-    "@angular/platform-browser": "16.0.0",
-    "@angular/platform-browser-dynamic": "16.0.0",
-    "@angular/router": "16.0.0",
-    "@ngrx/effects": "15.3.0",
-    "@ngrx/router-store": "15.3.0",
-    "@ngrx/store": "15.3.0",
-    "@ngrx/store-devtools": "15.3.0",
+    "@angular/animations": "17.0.2",
+    "@angular/cdk": "17.0.0",
+    "@angular/common": "17.0.2",
+    "@angular/compiler": "17.0.2",
+    "@angular/core": "17.0.2",
+    "@angular/forms": "17.0.2",
+    "@angular/material": "^16.2.0",
+    "@angular/platform-browser": "17.0.2",
+    "@angular/platform-browser-dynamic": "17.0.2",
+    "@angular/router": "17.0.2",
+    "@ngrx/effects": "16.0.1",
+    "@ngrx/router-store": "16.0.1",
+    "@ngrx/store": "16.0.1",
+    "@ngrx/store-devtools": "16.0.1",
     "animate.css": "^3.7.0",
     "bootstrap": "^5.0.1",
     "core-js": "3.19.0",
     "rxjs": "7.8.1",
     "rxjs-compat": "6.0.0",
     "tslib": "^2.4.0",
-    "zone.js": "0.13.0",
-    "@nx/angular": "16.1.3"
+    "zone.js": "0.14.2",
+    "@nx/angular": "17.1.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "16.0.0",
-    "@angular-devkit/core": "16.0.0",
-    "@angular-devkit/schematics": "16.0.0",
-    "@angular/cli": "~16.0.0",
-    "@angular/compiler-cli": "16.0.0",
-    "@angular/language-service": "16.0.0",
-    "@ngrx/schematics": "15.3.0",
-    "@nx/workspace": "16.1.3",
-    "@schematics/angular": "16.0.0",
+    "@angular-devkit/build-angular": "17.0.0",
+    "@angular-devkit/core": "17.0.0",
+    "@angular-devkit/schematics": "17.0.0",
+    "@angular/cli": "~17.0.0",
+    "@angular/compiler-cli": "17.0.2",
+    "@angular/language-service": "17.0.2",
+    "@ngrx/schematics": "16.0.1",
+    "@nx/workspace": "17.1.2",
+    "@schematics/angular": "17.0.0",
     "@types/jasmine": "4.0.3",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "14.14.33",
@@ -76,9 +76,9 @@
     "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-jasmine": "5.1.0",
     "karma-jasmine-html-reporter": "2.0.0",
-    "ng-packagr": "16.0.1",
-    "nx": "16.1.3",
-    "nx-cloud": "16.0.5",
+    "ng-packagr": "17.0.0",
+    "nx": "17.1.2",
+    "nx-cloud": "16.5.2",
     "postcss": "^8.4.14",
     "postcss-import": "14.1.0",
     "postcss-preset-env": "7.5.0",
@@ -87,6 +87,6 @@
     "protractor": "7.0.0",
     "ts-node": "10.9.1",
     "tslint": "6.1.3",
-    "typescript": "5.0.4"
+    "typescript": "5.2.2"
   }
 }
