@@ -1,30 +1,47 @@
-import { Component, OnInit } from '@angular/core';
-import { CarouselData } from '../app.component';
-import { Router, ActivatedRoute } from '@angular/router';
-import { tap } from 'node_modules/rxjs/operators';
-import { SlidesOutputData, OwlOptions } from 'ngx-owl-carousel-o';
+import { Component, OnInit, WritableSignal, signal } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { Router, ActivatedRoute, RouterOutlet, RouterLink } from '@angular/router';
+import { tap } from 'rxjs/operators';
+import { SlidesOutputData, OwlOptions, CarouselModule } from 'ngx-owl-carousel-o';
+
+interface CarouselData {
+  id: string;
+  text: string;
+  dataMerge?: number;
+  width: number;
+  dotContent?: string;
+  src?: string;
+  dataHash?: string;
+}
+
+
 @Component({
-  selector: 'owl-carousel-libdemo-home',
+  selector: 'app-home',
+  standalone: true,
+  imports: [CommonModule, CarouselModule, RouterOutlet, RouterLink],
   templateUrl: './home.component.html',
-  styleUrls: ['./home.component.sass']
+  styleUrl: './home.component.scss'
 })
 export class HomeComponent implements OnInit {
-  carouselData: CarouselData[] = [
-    { id: 'slide-1', text: 'Slide 1 HM', dataMerge: 2, width: 300, dotContent: 'text1'},
-    { id: 'slide-2', text: 'Slide 2 HM', dataMerge: 1, width: 500, dotContent: 'text2'},
-    { id: 'slide-3', text: 'Slide 3 HM', dataMerge: 3, dotContent: 'text3'},
-    { id: 'slide-4', text: 'Slide 4 HM', width: 450, dotContent: 'text4'},
-    { id: 'slide-5', text: 'Slide 5 HM', dataMerge: 2, dotContent: 'text5'},
-    { id: 'slide-6', text: 'Slide 6', dotContent: 'text5'},
+  carouselData: WritableSignal<CarouselData[]> = signal([
+    { id: 'slide-1', text: 'Slide 1 HM', dataMerge: 2, width: 300, dotContent: 'text1' },
+    { id: 'slide-2', text: 'Slide 2 HM', dataMerge: 1, width: 500, dotContent: 'text2' },
+    { id: 'slide-3', text: 'Slide 3 HM', dataMerge: 3, width: 500, dotContent: 'text3' },
+    { id: 'slide-4', text: 'Slide 4 HM', width: 450, dotContent: 'text4' },
+    { id: 'slide-5', text: 'Slide 5 HM', dataMerge: 2, width: 500, dotContent: 'text5' },
+    { id: 'slide-6', text: 'Slide 6', width: 500, dotContent: 'text5' },
+    { id: 'slide-7', text: 'Slide 7', width: 500, dotContent: 'text6' },
+    { id: 'slide-8', text: 'Slide 8', width: 500, dotContent: 'text8' },
     // { id: 'slide-7', text: 'Slide 7', dotContent: 'text5'},
     // { id: 'slide-8', text: 'Slide 8', dotContent: 'text5'},
     // { id: 'slide-9', text: 'Slide 9', dotContent: 'text5'},
     // { id: 'slide-10', text: 'Slide 10', dotContent: 'text5'},
-  ];
+  ]);
 
   customOptions: OwlOptions = {
-    autoWidth: true,
-    loop: true,
+    // autoWidth: true,
+    loop: false,
     // items: '10',
     // margin: 10,
     // slideBy: 'page',
@@ -32,8 +49,9 @@ export class HomeComponent implements OnInit {
     // autoplay: true,
     // autoplayTimeout: 5000,
     // autoplayHoverPause: true,
-		// autoplaySpeed: 4000,
+    // autoplaySpeed: 4000,
     dotsSpeed: 500,
+    rewind: false,
     // dots: false,
     // dotsData: true,
     // mouseDrag: false,
@@ -48,6 +66,7 @@ export class HomeComponent implements OnInit {
     // rtl: true,
     // startPosition: 1,
     // navText: [ '<i class=fa-chevron-left>left</i>', '<i class=fa-chevron-right>right</i>' ],
+    slideBy: 'page',
     responsive: {
       0: {
         items: 1
@@ -63,55 +82,59 @@ export class HomeComponent implements OnInit {
     nav: true
   }
 
-  currentUrl: any;
-  fragment: string;
+  currentUrl: WritableSignal<string> = signal('');
+  fragment: WritableSignal<string | null> = signal('');
 
-  activeSlides: SlidesOutputData;
+  activeSlides: WritableSignal<SlidesOutputData | null> = signal(null);
 
   constructor(private route: ActivatedRoute,
-              private router: Router) { }
+    private router: Router) { }
 
   ngOnInit() {
     // console.log(this.route.pathFromRoot);
     this.route.fragment.pipe(
-      tap(fragment => this.fragment = fragment),
-      // tap(() => console.log(this.fragment))
+      tap(fragment => this.fragment.set(fragment)),
+      // tap(() => console.log('this.fragment', this.fragment()))
     ).subscribe(
-      () => {}
+      () => { }
     );
 
     this.route.url.pipe(
-      tap(url => this.currentUrl = url[0].path),
-      // tap(() => console.log(this.currentUrl))
+      tap(url => this.currentUrl.set(url[0].path)),
+      // tap(() => console.log('this.currentUrl', this.currentUrl()))
     ).subscribe(
-      () => {}
+      () => { }
     )
 
   }
 
   moveToSS() {
-    this.router.navigate(['/' + this.currentUrl], {fragment: 'second-section'});
+    this.router.navigate(['/' + this.currentUrl()], { fragment: 'second-section' });
   }
 
   getPassedData(data: any) {
-    this.activeSlides = data;
+    this.activeSlides.set(data);
     console.log('HomeComponent');
-    console.log(this.activeSlides);
+    console.log(this.activeSlides());
   }
 
   getChangeData(data: any) {
-    this.activeSlides = data;
+    this.activeSlides.set(data);
     console.log('HomeComponent -> change');
     console.log(data);
   }
 
   getChangedData(data: any) {
-    this.activeSlides = data;
+    this.activeSlides.set(data);
     console.log('HomeComponent -> changed');
     console.log(data);
   }
   removeLastSlide() {
-    this.carouselData.splice(-1, 1);
+    this.carouselData.update((data) => {
+      const res = [...data]
+      res.splice(-1, 1)
+      return res
+    });
   }
 
   carouselChanged(evt: SlidesOutputData) {
