@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { DoubledCarouselComponent } from './doubled-carousel.component';
 
@@ -6,14 +6,12 @@ describe('DoubledCarouselComponent', () => {
   let component: DoubledCarouselComponent;
   let fixture: ComponentFixture<DoubledCarouselComponent>;
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ DoubledCarouselComponent ]
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [DoubledCarouselComponent]
     })
     .compileComponents();
-  }));
-
-  beforeEach(() => {
+    
     fixture = TestBed.createComponent(DoubledCarouselComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
