@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PresentComponent } from './present.component';
 
@@ -6,14 +6,12 @@ describe('PresentComponent', () => {
   let component: PresentComponent;
   let fixture: ComponentFixture<PresentComponent>;
 
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PresentComponent ]
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [PresentComponent]
     })
     .compileComponents();
-  }));
-
-  beforeEach(() => {
+    
     fixture = TestBed.createComponent(PresentComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
