@@ -2,7 +2,7 @@
 
 // import 'core-js/es7/reflect';
 import 'zone.js';
-import 'zone.js/dist/zone-testing';
+import 'zone.js/testing';
 import { getTestBed } from '@angular/core/testing';
 import {
   BrowserDynamicTestingModule,
